server/fastapi_backend/requirements.txt
fastapi
uvicorn
pymongo
python-jose
passlib[bcrypt]
pydantic
